+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+1.2.1
+=====
+Added in support for additional parameters for flutter build runs (flavours, device ids and
+generic additional parameters).
+
 1.1.2
 =======
 Added in override support for which file to actually run